+++ everett/__init__.py
@@ -28,6 +28,6 @@ __author__ = 'Will Kahn-Greene'
 __email__ = 'willkg@mozilla.com'
 
 # yyyymmdd
-__releasedate__ = ''
+__releasedate__ = '20161108'
 # x.y or x.y.dev
-__version__ = '0.5.dev'
+__version__ = '0.5'